Restructuring EquippableGear
This was a very fun commit! I imported and wrote some more planning work in the Obsidian library, made some design decisions and implemented a basic form of the whole EquippableGear system and all that. My notes file was 3-2 Progress.

What I plan to do next is go and make some simple animations for, and implement some of the fishing rod functionality.

:camel: