FishingRod Casting Part I :fishing_pole_and_fish:
This was quite a big commit, as I was getting the first bits of casting the fishing rod out. This involves a few kinematic equations, ConcreteGearState states, some raycasting, and gizmo  visualization. I've created an algorithm to make a list of test points, simulating if a collision occurs alongside the desired kinematic casting path. If so, we'll create a list of test points that ends with the first pair of test points that experienced the collision, and returns the point of contact as well.

The next steps are working on the states in which the line is cast and where you can reel the line back (semi-instantly, unsimulated). After this is done, I can start doing neat stuff with the actual fishing minigame / mechanic.

:camel: