(Incomplete) Placing Bobber Preview
I may resume work on my laptop so this commit is to transfer it. I've gotten some work done and completed the kinematic equations, aiming, and more. Now, however, I need to place the new bobber preview object (and have it stay) so that I can move the actual bobber to that position in the next state.

:camel: