First Casting Approach is Functional!
I've added a BobberScript to the bobber and created a little function to launch it all well and stuff like that. I'm so glad! There's a lot of refining I could do (adding a transform/parent GameObject to hold the Bobber to attach to /detach from), but my initial approach is working. Huzzah! Now to register that we're fishing and make the actual fishing action!

:camel: