Upgraded to Visual Studio 2022 + More Lighting!
Not a significant amount different script-wise or innovation-wise here, just some work for my GMD class and all that. I tried to come up with a decent organization for lighting in the Hierarchy, we'll have to see if that rules or not.
As well, I have regenerated project files to include Player data for VS2022, which may have implications I'm not aware of. I won't sweat it though!

:camel: