Implementing Non-Kinematic Bobber (PART 1)
I figured that it would be cooler, more fun, and more of a challenge to implement a non-kinematic bobber to be launched. There's still a lot of work that I have to do if I'm to get this done (significantly less than if I was to try a kinematic system full of calculations that wouldn't allow for bouncing, banking, ramping action).

:camel: