Removing Kinematic TestPoint System
Since I decided against using a kinematic system, I have little use for this currently. Thankfully, I still have the notes and commits of this work before in the event I have to use it again for some reason. In any case, this is the first step to a fully non-kinematic bobber system.

:camel: