Attempting Bobber Floating, Gathering Analytics
I'm trying to make the bobber float on a given body of water, but I'm trying to find a decent way to do so. There's many different approaches but I want to try and keep it non-kinematic for the moment. That said, I'm not done with this, I'm just marking my changes for now so I can pick up after today's Overwatch League matches.

:camel: