Documenting before Next Session
The whole thing here is that Rigidbodies and parents tend to disagree about position, rotation, scale, all the little things in life... and it ruined my little reeling-in process. When I'm back at the first state and I need to re-parent the Bobber to the fishing rod, eldritch mathematical horrors change my values, even with lossy scaling applied.

That said, next session I'm going to learn and implement joints. Onwards and upwards!

:camel: