Joint Work Day 1 Work
I did my initial attempts at implementing joints today, which are documented. There is plenty more work to be done, like actually getting my  fishing rod and joint system working, and working with my states and things like that. I'll have to see how to do that, but that'll be work for tomorrow (or later today since I forgot to commit this last night!)
:camel: