Upgrading Casting Functionality w/ Physics
I have an OW scrimmage to coach so I will have to leave this here. At present, I'm working to attach and detach the bobber. Now, however, I've restored all of the original functionality of the project prior to this re-prioritization. In any case, my doc file says it all.

:camel: