Got cast and re-casting working with FixedJoint
Simply, I am now making the dynamic Rigidbody for the bobber actually JOINTED to the rod when it's inactive. This FINALLY seems to close out the issue I had with parenting and crazy things, so I can continue on.

Next, I'm thinking about getting the entire system fully loop-able and with the physics tuned up better.

:camel: